requirements.txt
vtk
trimesh
numpy
flask
flask-cors
torch
torchvision
Pillow
opencv-python
ultralytics
